backend/requirements.txt
fastapi
uvicorn
websockets
pytest
pytest-asyncio
google-adk